Tactile-only FB in calibration
Calibration was changed so as to provide tactile only feedback, analysis window was shortened to one sec, a dead-zone has been added in order to prevent selection with uncertain values. There's still problems: accuracy is lowered by lack of frequency filter in Simulink model (unsure as to how they were removed), sequence of operations on experiment start is byzantine, timing in the interface is way off. 
Beside adding freq filters, code should be streamlined somewhat and changed so that during use no intention estimation occurs during countdown, but it starts afterwards, then continues until a decision can be taken with a modicum of certainty.