Working version, but still no luck with tests
The code should already be working as is, but apparently I am now unable to generate motor imagery on command. Testing with older setups, it would seem that both error potentials and visual-related activity are produced and can be detected normally, so the idea now is to change the interface and use a classifier that takes into account both motor imagery estimations and error potentials and combine them to produce an estimation of actual intention. This requires essentially scrapping the whole interface and starting anew (yeah!), so I shall go to work now.