Added dynamic iteration number
Code has been modified so that the number of iterations per trial is not fixed anymore. Instead, the software will keep proposing iterations until it is 99% sure of current selection (will see if this actually works).
Furthermore, calibration has been instructed to stop when accuracy reaches 90% over the last 15 iterations.