System works (probably) with unorthodox classifier
Classifier proposed in the paper has been substituted with a rather unconventional alternative:  new classifier is trained on both training and testing data sets, assuming both possible label configurations for the testing set. The classifier resulting in the highest accuracy determines the correct labeling of the test set. This is only implemented in an offline version, need to modify it so that it works online, too.