I am a sausage: forgot to modify freq filters
So, first experiment on subject was basically screwed: all data is processed with the same set of freq filters (namely, a bandpass between 0.1 and 10 Hz), meaning that essentially NO power in the relevant band is left to decode motor imagery correctly. After correcting for this offline, trial accuracy for S01 reaches 0.7something, up from <0.6. Need to modify code to correctly process both errP and MI data (need to modify Simulink model, for that)